Revert "Refactor: modularize bot commands"
